bug fix
in previous version, it did not distinguish whether log status is 'D'
or not.
and if error exist, printed out redundant outputs as standard io.
they must be printed as log. in the future, I plan to implement 
log function.
and if error exist, it did not distinguish whether -ignore-err is
indicated or not.
in this version, above problems are fixed. but log function is not
implemented yet.